Move marker handling to Actor
